[c++/empty cells] Empty cell information is now displayed with "" instead of "_openxlsx_NA_" as before. Previously this variable was used to help various `wbWorkbook` functions to distinguish between empty cells, missing cells and cells to be ignored. Fortunately, this is no longer needed. Either a cell contains useful information or it can be ignored. Empty strings are written as `<is><t></t></is>` or they are still sharedstrings. Row and column information cannot be empty, similarly openxml has no empty attributes (as far as I know).
This relieves us of the time-consuming burden of initializing each field of cc with the long string "_openxlsx_NA_" (it was intentionally that long to prevent someone from randomly selecting it in an xlsx file). However, since this string was used fundamentally in several places, this change should be monitored closely. Especially for files that are known to be strange (missing lines, missing data, empty strings, or uninitialized cells).

There is another string called "_openxlsx_NA". This string has another use that is still required, and writing this string from openxlsx2 will still corrupt the workbook as indicated in the description.