No drawings dir (#311)
* avoid adding drawings and vmlDrawings to every sheet created

* helper function to read only the content/children of a node. <xml><a/></xml> will return <a/>.
Provide no_declaration function to write files without declaration. A certain spreadsheet software does not seem to care at all.

* handle empty vml

* Revert "helper function to read only the content/children of a node. <xml><a/></xml> will return <a/>."

This reverts commit b0d385555ffcf114085a3ff67f3b1cdd2b87221e.

* update tests

* fix test, but this clone function is broken anyways

* correct fix for clone_worksheet with tables

* fix bad rebase

* fix tests after rebase
This is a partial reland. For vml writing, if the sheet already contains vml (a button etc.) you could not add a comment to it.

* fix vml_comment

* restore comment logic
this allows writing multiple comments as well as writing comments to sheets with other vml objects

* use drawing and vmlDrawing id

* fix previous commit

* fix the fix

* update NEWS.md

* fix remove_worksheet
for now we keep the drawings references as is. if worksheet 2 references drawings2.xml, it will not change to drawings1.xml, if sheet 1 is removed and if drawings1.xml becomes available. this avoids changing every drawings reference in all worksheets numbers higher than the removed one.