[write] use custom `as_character()`
The bottleneck when writing data, is the conversion from numerics to characters. We use as.character() which creates deferred string conversions, aka ALTREP lazily string conversions that materialize only if explicity called. In our case, when add column names in the colnames example, or when passing the object to Rcpp. There is no real way around this, because at one point we have to convert the logical, numeric or date object into an R object. Still, we could try to find quicker ways to avoid as.character() as in the as_character() function. Obviously the question remains, do we want to risk bad conversions for marginal speed gains. This will bring anoter 20 ms or so, when writing a 2000x100 data frame. There might be other tricks like searching for uniques and converting these in bulk. Ultimately though, I'm not yet happy with this.