[read+write] consolidate `na.strings` and `na.numbers` into `na` (#1499)
* [read+write] consolidate `na.strings` and `na.numbers` into `na`

`na.strings` is a relict from openxlsx and should have been replaced with na_strings a long time ago. But it was not and now there is `na_strings()` and we cannot have `na_strings = na_strings()`. Thus this PR is the attempt to use the short form `na` instead of `na.strings`

* [write] reorder options

Co-authored-by: olivroy <52606734+olivroy@users.noreply.github.com>

* [tests] more tests for `na`

* [write] add special for NULL

* [misc] update NEWS

* [read] cleanup

* [read] avoid conversion error on missing dates

* [write] cleanup

---------

Co-authored-by: olivroy <52606734+olivroy@users.noreply.github.com>